requirements.txt
requests>=2.32.3
selenium>=4.29.0
python-dotenv>=1.0.1
cloudscraper>=1.2.71
seleniumbase>=4.37.7
playwright>=1.52.0
discord.py>=2.4.0

